+++ build_runner/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bug fix: reject incorrect builder config that has duplicate builder keys.
 - Bug fix: reject invalid builder factories in `build.yaml`.
 - Bug fix: correctly filter served or `--output` post process output.
+- Bug fix: `serve` mode only serves the specified package subdirectories.
 
 ## 2.15.2